+++ kitty/options/to-c-generated.h
@@ -942,6 +942,19 @@ convert_from_opts_allow_hyperlinks(PyObject *py_opts, Options *opts) {
 }
 
 static void
+convert_from_python_menu_map(PyObject *val, Options *opts) {
+    menu_map(val, opts);
+}
+
+static void
+convert_from_opts_menu_map(PyObject *py_opts, Options *opts) {
+    PyObject *ret = PyObject_GetAttrString(py_opts, "menu_map");
+    if (ret == NULL) return;
+    convert_from_python_menu_map(ret, opts);
+    Py_DECREF(ret);
+}
+
+static void
 convert_from_python_wayland_titlebar_color(PyObject *val, Options *opts) {
     opts->wayland_titlebar_color = PyLong_AsUnsignedLong(val);
 }
@@ -1229,6 +1242,8 @@ convert_opts_from_python_opts(PyObject *py_opts, Options *opts) {
     convert_from_opts_close_on_child_death(py_opts, opts);
     if (PyErr_Occurred()) return false;
     convert_from_opts_allow_hyperlinks(py_opts, opts);
+    if (PyErr_Occurred()) return false;
+    convert_from_opts_menu_map(py_opts, opts);
     if (PyErr_Occurred()) return false;
     convert_from_opts_wayland_titlebar_color(py_opts, opts);
     if (PyErr_Occurred()) return false;