+++ kitty/state.h
@@ -23,6 +23,12 @@ typedef struct {
 
 typedef enum AdjustmentUnit { POINT = 0, PERCENT = 1, PIXEL = 2 } AdjustmentUnit;
 
+struct MenuItem {
+    const char* *location;
+    size_t location_count;
+    const char *definition;
+};
+
 typedef struct {
     monotonic_t visual_bell_duration, cursor_blink_interval, cursor_stop_blinking_after, mouse_hide_wait, click_interval;
     double wheel_scroll_multiplier, touch_scroll_multiplier;
@@ -92,6 +98,7 @@ typedef struct {
     int background_blur;
     long macos_titlebar_color;
     unsigned long wayland_titlebar_color;
+    struct { struct MenuItem *entries; size_t count; } global_menu;
 } Options;
 
 typedef struct WindowLogoRenderData {
@@ -343,6 +350,7 @@ typedef enum {
     HIDE_OTHERS,
     MINIMIZE,
     QUIT,
+    USER_MENU_ACTION,
 
     NUM_COCOA_PENDING_ACTIONS
 } CocoaPendingAction;